Add support for VSCode
With this change you should be able to use VSCode instead of eclipse for programming. Simply install VSCode, install the Java extension (install OpenJDK if it later prompts you to do so) and then open this folder in VSCode. You can build with Shift+Alt+B. Deploying will still need to be done via eclipse, but primary development can be a lot easier now.

Note: The pom.xml file is the key that makes this work, feel free to share this with other teams that might want to use VSCode (I found examples from previous years, but no one has posted one for this year).