Editing EncoderDrive + Commands + Subsystems
For our situation, it is best to use the Relative setting for the
encoders. We have switched from using Absolute because reading Absolute
values are for a slightly different purpose. Our understanding of the
encoder code thus far is that our robot will coast slightly past our
target goal. At the moment, there is a simple correction multiplier that
will get fine-tuned as soon as we are able. The Gripper commands were
remade to actually make sense and be usable. Other edits to fix some
null references were made. Also, extra sorry to Gurman for renaming the
subsystems xd.