Xbox controller for 2nd driver and auto routines
The code works for turning perfectly now. The encoder is working great
with a 5% error which is pretty good. We're using an Xbox controller so
the second driver can now Control the gripper and elevator yea screw you
Gurman. Fixed naming and also added overloading to code meaning you can
addSequential to EncoderDrive and you can set a timeout when you want
to.
ToDo: Code all the routines. Right now, do crosslines then when we fix
the mechanism we can start doing cube on scale and switch