Lots of stuff
Everything that includes setting motor speed or directly contacting the
cantalon or setting the values will be in a subsytem now. A command
should never have a number in it unless it's a timeout. If the other
programmers have a problem with this, discuss it with me because I might
be wrong. In driving, when the elevator goes above a certain height,
slow down driving as to not tilt. I added the auto commands for grip
motors and pneumatics but I kinda forgot how to make it so that when you
call the command, you can access one of two thingsf which are intake or
outtake and open pneumatics or close. Maybe use a switch or enum? Hope I
remember eventually