Auto format + New gripper control + Elevator tuning
Firstly, our code looked disgusting so I ran the autoformatter to
actually indent our code. Should probably be doing that more regularly.
Gurman and Terrance were complaining about not being able to pick up the
cubes well from the corners, so I added controls to run the gripper
motors independantly to turn the cube for better grip. The elevator will
also automatically close the gripper at a certain height based off of
the encoder. We may need to zero the encoder at a different time.
Currently, it zeros when the robot is initialized, but that isn't
assured to be the elevator's lowest point.