All autonomous routines I could think of
These are all the autonomous routines I can think of. The switches are
made so that after dropping cube off, it crosses the line. I don't know
if we'll ever do scale from middle pos. If we can do scale, then we are
top prio so it should be ok. When we're forced in the middle pos, we can
place a cube on the left or right switch side and/or cross line